requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
requests==2.25.1
urllib3==1.26.5

pandas~=1.2.4
TA-lib~=0.4.20
numpy~=1.20.3
alpaca-trade-api==1.2.1
finnhub-python==2.3.0
loguru~=0.5.3
schedule~=1.1.0